fix: address copilot PR#3 review — division-by-zero guards, breakdown merge, doc fixes
- Empty scorer arrays: guard against div-by-zero in PHP/Python/Rust/TS
- CompositeScorer: merge breakdowns from all scorers (was first-only)
- PHP: replace PECL yaml_emit with symfony/yaml for portability
- PHP: guard ConformanceTest yaml_parse with function_exists
- Rust: fix serde_yaml Value::Number type conversion
- Docs: fix Go/PHP/Python guides referencing non-existent APIs
- Add regression tests: TS (5), PHP (3), Python (2), Rust (2)

+++ rs/src/session.rs
@@ -277,7 +277,7 @@ where
             // Collect observations for next turn
             context.insert(
                 "step".into(),
-                serde_yaml::Value::Number(self.step.into()),
+                serde_yaml::Value::Number((self.step as u64).into()),
             );
 
             results.push(result);
